pyproject.toml
[project]
name = "interactive-feedback-mcp"
version = "0.1.0"
description = "MCP server for interactive user feedback and command execution in AI-assisted development, by Fábio Ferreira."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "psutil>=7.0.0",
    "pyside6>=6.8.2.1",
    "pillow",
]
